Fix invalid swatch regex crashing app
As opposed to 88372ec8458cf5e26efc653e75583f53784f3dc1, this actually
fixes the issue by escaping the regex before attempting to generate the
fallback. It wasn't doing this before which caused the exact issue that
was being fixed, but with the slashes and flags on either end instead.